+++ src/main/java/pl/kupniewski/application/simulation/RestaurantSimulator.java
@@ -1,7 +1,7 @@
 package pl.kupniewski.application.simulation;
 
 import org.springframework.stereotype.Component;
-import pl.kupniewski.api.restaurant.CompleteOrderByRestaurant;
+import pl.kupniewski.application.usecases.deliver.CompleteOrderByRestaurant;
 import pl.kupniewski.application.usecases.deliver.OrderReadyToDeliver;
 import pl.kupniewski.application.usecases.order.Order;
